Move data processing functions to IndicatorStats utils
