+++ components/IndicatorStat/utils/data.ts
