+++ components/IndicatorStat/utils/index.ts
